StatementResource implementation 1) Make optional properties on StatementEntity nullable 2) Set StatementEntity result and foreign key properties e.g. contententryuid, cbuid etc 3) Store Verb and and verb langmap entities in the database.

+++ core/src/commonMain/kotlin/com/ustadmobile/core/domain/xapi/model/Result.kt
@@ -17,9 +17,9 @@ data class Result(
     @Serializable
     data class Score(
         var scaled: Float,
-        var raw: Long,
-        var min: Long,
-        var max: Long
+        var raw: Float,
+        var min: Float,
+        var max: Float,
     )
 
 }
